Agile-SD : TCP congestion controller
Agile-SD is a new loss-based and RTT-independent TCP[1] congestion control algorithm designed to support high-speed and Short-Distance (SD) networks.
It mainly contributes the agility factor mechanism, which allows Agile-SD to deal with small buffer sizes while reducing its sensitivity to packet loss.
Due to the use of this mechanism, Agile-SD improves the throughput of TCP up to 50% compared to Cubic-TCP and Compound-TCP.
Its performance was evaluated using the well-known NS2 simulator to measure the average throughput, loss ratio and fairness.

Agile-SD[2] is designed and implemented by Alrshah et al. (2015) as a Linux kernel module in the real Linux operating system (openSUSE 42.1 Leap)
which is implemented at the Network, Parallel and Distributed Computing Laboratory, A2.10, second floor, Block A, Faculty of
Computer Science and Information Technology, Universiti Putra Malaysia, over real PCs connected to the Internet for daily uses and evaluation purposes.
The main motivation behind Agile-SD is to support the short-distance networks such as local area networks
and data center networks in order to increase the bandwidth utilization over high-speed networks.
Moreover, Agile-SD is introduced in support of the open source community, where it is can be used under the OpenGL agreement.

Authored by : Mohamed A. Alrshah
patch taken for https://patchwork.kernel.org/patch/9874023

Signed-off-by: Haseo97 <jawgattus@gmail.com>

And set it as default.

	modified:   arch/arm64/configs/lavender-perf_defconfig
	modified:   net/ipv4/Kconfig
	modified:   net/ipv4/Makefile
	new file:   net/ipv4/tcp_agilesd.c